Revert "User register (#52)"
This reverts commit 07c09fe50ba7fa9c43d3a9f0ececbb7db18557d1.
