+++ .gitignore
@@ -5,4 +5,3 @@ myvenv
 db.sqlite3
 /static
 .DS_Store
-/TODO
