+++ car_auction_service/car_auction_service/settings.py
@@ -141,7 +141,6 @@ LOGIN_REDIRECT_URL = "/cars"
 LOGOUT_REDIRECT_URL = "/cars"
 
 
-
 INTERNAL_IPS = [
     "127.0.0.1"
 ]
