+++ car_auction_service/cars/forms.py
@@ -17,3 +17,18 @@ class ImageForm(forms.ModelForm):
     class Meta:
         model = Image
         fields = ["image"]
+
+
+class UserRegistrationForm(forms.ModelForm):
+    password = forms.CharField(label="Password", widget=forms.PasswordInput)
+    password2 = forms.CharField(label="Repeat password", widget=forms.PasswordInput)
+
+    class Meta:
+        model = User
+        fields = ('username', 'first_name', 'email')
+
+    def clean_password2(self):
+        cd = self.cleaned_data
+        if cd['password'] != cd['password2']:
+            raise forms.ValidationError('Passwrods are not the same!')
+        return cd
