+++ car_auction_service/templates/users/register.html
@@ -1,15 +1,10 @@
-{% extends "base.html" %}
-
-{% block title %}
-	{{ User register form }}
-{% endblock %}
-
-	<h1>Create Account</h1>
-	<p>Fill below form to create an account</p>
-	<form action="." method="POST">
-		{{% csrf_token %}}
-		{{ user_form.as_p }}
-		<p><input type="submit" value="Create an account"></p>
-	</form>
-
-{% endblock %}
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8">
+		<title>Title</title>
+	</head>
+	<body>
+	
+	</body>
+</html>
