+++ car_auction_service/users/forms.py
@@ -2,15 +2,15 @@ from django.contrib.auth.models import User
 from django import forms
 
 class UserRegistrationForm(forms.ModelForm):
-    password = forms.CharField(label="Password", widget=forms.PasswordInput)
-    password2 = forms.CharField(label="Repeat password", widget=forms.PasswordInput)
+    password = forms.CharField(label='Hasło', widget=forms.PasswordInput)
+    password2 = forms.CharField(label='Powtórz hasło', widgets=forms.PasswordInput)
 
     class Meta:
         model = User
-        fields = ('username', 'first_name', 'email')
+        fields = ('Username', 'first_name', 'emial')
 
     def clean_password2(self):
         cd = self.cleaned_data
         if cd['password'] != cd['password2']:
-            raise forms.ValidationError('Passwordsare not the same!')
+            raise forms.ValidationError('Hasła nie są identyczne/')
         return cd['password2']
