+++ car_auction_service/users/urls.py
@@ -21,7 +21,6 @@ app_name = 'users'
 
 urlpatterns = [
     path("accounts/", include("django.contrib.auth.urls")),
-    path("accounts/register", views.register, name='register')
     # path('', TemplateView.as_view(template_name='home.html'), name='home')
 
 ]
