+++ car_auction_service/users/views.py
@@ -1,18 +1,3 @@
 from django.shortcuts import render
 
-from .forms import UserRegistrationForm
-
 # Create your views here.
-
-def register(request):
-    if request.method == "POST":
-        user_form = UserRegistrationForm(request.POST)
-        if user_form.is_valid():
-            new_user = user_form.save(commit=False)
-            new_user.set_password(user_form.cleaned_data['password'])
-            new_user.save()
-            return render(request, 'account/register_done.html', {'new_user':new_user})
-    else:
-        user_form = UserRegistrationForm()
-    return render(request, 'account/register.html', {'user_form':user_form})
-    #TODO: add templates register_done and register
